pyproject.toml
[project]
name = "tax-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.73.1",
    "mcp[cli]>=1.27.0",
    "ollama>=0.6.1",
    "pypdf>=6.10.2",
    "pyyaml>=6.0.3",
    "qwen-agent[mcp]>=0.0.34",
    "fastmcp>=2.0.0",
    "numpy>=2.4.4",
    "soundfile>=0.13.1",
    "python-dateutil>=2.9.0.post0",
]
